Adjust Auto Deploy to follow main drone's yaw

+++ gcs-server/domain/src/main/java/com/gcs/domain/context/dto/AgentContextDto.java
@@ -1,21 +1,16 @@
 package com.gcs.domain.context.dto;
 
-import com.gcs.domain.agent.Agent;
 import com.gcs.domain.agent.dto.AgentDto;
 import com.gcs.domain.context.AgentContext;
 import com.gcs.domain.context.constants.AgentContextConnection;
-import lombok.val;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Stream;
 
 public class AgentContextDto implements AgentContext {
     private AgentContextConnection connection;
     private final Map<Integer, AgentDto> agents;
-    private Integer rotation;
+    private Float rotation;
 
     public AgentContextDto() {
         agents = new HashMap<>();
@@ -35,11 +30,11 @@ public class AgentContextDto implements AgentContext {
         this.connection = connection;
     }
 
-    public void setRotation(int rotation){
+    public void setRotation(float rotation){
         this.rotation = rotation;
     }
 
-    public int getRotation(){
+    public Float getRotation(){
         return rotation;
     }
